Pull out ObjectProperty to sit alongside data class.
